Create ReadWriteFile.py
A text file contains some text, we need to know how often (or how rare) each letter appears in the text. Your task is to write a program which asks the user for the input file's name; reads the file (if possible) and counts all the Latin letters (lower- and upper-case letters are treated as equal); prints a sorted histogram based on the characters' frequency (the bigger counter should be presented first) (only non-zero counts should be presented); and also the histogram should be sent to a file.